Adição dos gráficos

+++ ProjectFacul/ProjectFacul.ProjetoWeb.Api/Model/DadosTemperatura.cs
@@ -4,7 +4,7 @@ namespace ProjectFacul.ProjetoWeb.Api.Model
 {
     public class DadosTemperatura
     {
-        public int Temperatura { get; set; }
+        public float Temperatura { get; set; }
         public DateTime Data { get; set; }
     }
 }
